src_rx/requirements.txt
influxdb-client[ciso]>=1.37
pyserial>=3.5
